LayoutTests/http/tests/site-isolation/color-input-cross-origin-iframe.html

<!DOCTYPE html><!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
<script src="/js-test-resources/ui-helper.js"></script>
</head>
<body>
<script>
description("Test that selecting a color from the color picker in a cross-origin iframe updates the input value when site isolation is enabled.");
jsTestIsAsync = true;

window.addEventListener("message", async (event) => {
    if (event.data.type === "result") {
        window.selectedValue = event.data.value;
        shouldBeEqualToString("selectedValue", "#ff0000");
        finishJSTest();
    }
});

async function runTest() {
    const iframe = document.getElementById("frame");
    const iframeRect = iframe.getBoundingClientRect();

    const clickX = iframeRect.left + 30;
    const clickY = iframeRect.top + 20;
    await UIHelper.activateFormControlAtPoint(clickX, clickY);

    // Select red (RGB: 1, 0, 0)
    await UIHelper.setSelectedColorForColorPicker(1, 0, 0);
}
</script>
<iframe id="frame" src="http://localhost:8000/site-isolation/resources/color-input-iframe.html" onload="runTest()"></iframe>
</body>
</html>
